+++ app/helpers/request_helper.py
@@ -4,9 +4,22 @@ from requests.exceptions import ConnectionError
 from app.helpers.exceptions import RetryableError
 
 
-def remote_call(url, json=None):
+def service_name(url=None):
     try:
-        logger.info("Calling service", request_url=url)
+        parts = url.split('/')
+        if 'responses' in parts:
+            return 'SDX_STORE'
+        elif 'sequence' in parts:
+            return 'SDX_SEQUENCE'
+    except AttributeError as e:
+        logger.error(e)
+
+
+def remote_call(url, json=None):
+    service = service_name(url)
+
+    try:
+        logger.info("Calling service", request_url=url, service=service)
         response = None
 
         if json:
@@ -24,22 +37,24 @@ def remote_call(url, json=None):
         raise RetryableError("Connection error")
 
 
-def response_ok(response):
+def response_ok(response, service_url=None):
+    service = service_name(service_url)
+
     if response is None:
         logger.error("No response from service")
         return False
     elif response.status_code == 200:
-        logger.info("Returned from service", request_url=response.url, status=response.status_code)
+        logger.info("Returned from service", request_url=response.url, status=response.status_code, service=service)
         return True
     else:
-        logger.error("Returned from service", request_url=response.url, status=response.status_code)
+        logger.error("Returned from service", request_url=response.url, status=response.status_code, service=service)
         return False
 
 
 def get_sequence_no():
     sequence_url = "{0}/json-sequence".format(SDX_SEQUENCE_URL)
     response = remote_call(sequence_url)
-    if not response_ok(response):
+    if not response_ok(response, sequence_url):
         return None
 
     result = response.json()
@@ -50,7 +65,7 @@ def get_doc_from_store(mongoid):
     store_url = "{0}/responses/{1}".format(SDX_STORE_URL, mongoid)
     response = remote_call(store_url)
 
-    if not response_ok(response):
+    if not response_ok(response, store_url):
         return None
 
     result = response.json()