+++ tests/test_request_helper.py
@@ -27,3 +27,23 @@ class TestRequestHelper(unittest.TestCase):
         response.status_code = 500
         result = request_helper.response_ok(response)
         self.assertEqual(result, False)
+
+    def test_service_name_return_responses(self):
+        url = "www.testing.test/responses/12345"
+        service = request_helper.service_name(url)
+        self.assertEqual(service, 'SDX_STORE')
+
+    def test_service_name_return_sequence(self):
+        url = "www.testing.test/sequence"
+        service = request_helper.service_name(url)
+        self.assertEqual(service, 'SDX_SEQUENCE')
+
+    def test_service_name_return_none(self):
+        url = "www.testing.test/test/12345"
+        service = request_helper.service_name(url)
+        self.assertEqual(service, None)
+
+    def test_service_name_url_none(self):
+        url = None
+        service = request_helper.service_name(url)
+        self.assertEqual(service, None)